Attach functionality to HTML buttons
Made js functions and attached them to buttons for resizing the
etch-a-sketch, changing to rainbow etch, and changing to black etch.
This allows the user to customize their etch-a-sketch experience and
means this etch-a-sketch project has all of it's major functions. The
js side of things is mostly completed, although the code to reset the
etch-a-sketch is repeated several times and may be created into its
own function.